lambdas/requirements.txt
aniso8601==7.0.0
boto3==1.18.44
botocore==1.21.44
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
jmespath==0.10.0
promise==2.3
psycopg2-binary==2.9.1
PyPika==0.48.8
python-dateutil==2.8.2
Rx==1.6.1
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
